pyproject.toml
[project]
name = "a2a-check"
version = "0.1.0"
description = "CLI tool to validate and inspect A2A Agent implementations"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Markus" }]
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "rich>=13.7",
  "pydantic>=2.8",
  "a2a-sdk>=0.3.1",
  "httpx-sse>=0.4.1",
]

[project.scripts]
a2a-check = "a2a_check.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_check"]
